+++ app/src/androidTest/java/org/simple/clinic/storage/migrations/Migration116AndroidTest.kt
@@ -6,7 +6,7 @@ import org.simple.clinic.assertColumns
 class Migration116AndroidTest : BaseDatabaseMigrationTest(115, 116) {
 
   @Test
-  fun smoking_should_be_added_to_medical_history_table() {
+  fun isSmoking_should_be_added_to_medical_history_table() {
     before.assertColumns("MedicalHistory", setOf(
         "uuid",
         "patientUuid",
@@ -33,7 +33,7 @@ class Migration116AndroidTest : BaseDatabaseMigrationTest(115, 116) {
         "hasHadStroke",
         "hasHadKidneyDisease",
         "hasDiabetes",
-        "smoking",
+        "isSmoking",
         "syncStatus",
         "createdAt",
         "updatedAt",