+++ app/src/main/java/org/simple/clinic/medicalhistory/OngoingMedicalHistoryEntry.kt
@@ -20,7 +20,7 @@ data class OngoingMedicalHistoryEntry(
     val isOnHypertensionTreatment: Answer = Unanswered,
     val isOnDiabetesTreatment: Answer = Unanswered,
     val hasDiabetes: Answer = Unanswered,
-    val smoking: Answer = Unanswered,
+    val isSmoking: Answer = Unanswered,
 ) : Parcelable {
 
   fun answerChanged(question: MedicalHistoryQuestion, answer: Answer): OngoingMedicalHistoryEntry {