+++ app/src/main/java/org/simple/clinic/medicalhistory/sync/MedicalHistoryPayload.kt
@@ -43,7 +43,8 @@ data class MedicalHistoryPayload(
     @Json(name = "hypertension")
     val hasHypertension: Answer?,
 
-    val smoking: Answer,
+    @Json(name = "smoking")
+    val isSmoking: Answer,
 
     @Json(name = "created_at")
     val createdAt: Instant,